refactor/#9: 주식 스크리너 위젯 리팩토링
